Move iterable handler and reader into own namespace
