+++ src/Reader/Iterable/FilterHandler/EqualsNull.php
@@ -2,22 +2,23 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Data\Reader\IterableFilterHandler;
+namespace Yiisoft\Data\Reader\Iterable\FilterHandler;
 
 use InvalidArgumentException;
+use Yiisoft\Arrays\ArrayHelper;
 use Yiisoft\Data\Reader\FilterAssertHelper;
-use Yiisoft\Data\Reader\IterableFilterHandlerInterface;
+use Yiisoft\Data\Reader\Iterable\IterableFilterHandlerInterface;
 
 use function count;
 
 /**
- * EqualsEmpty iterable filter handler checks that the item's field value is empty.
+ * EqualsNull iterable filter handler checks that the item's field value is null.
  */
-final class EqualsEmpty implements IterableFilterHandlerInterface
+final class EqualsNull implements IterableFilterHandlerInterface
 {
     public function getOperator(): string
     {
-        return \Yiisoft\Data\Reader\Filter\EqualsEmpty::getOperator();
+        return \Yiisoft\Data\Reader\Filter\EqualsNull::getOperator();
     }
 
     public function match(array|object $item, array $arguments, array $iterableFilterHandlers): bool
@@ -30,6 +31,6 @@ final class EqualsEmpty implements IterableFilterHandlerInterface
         FilterAssertHelper::assertFieldIsString($field);
 
         /** @var string $field */
-        return empty($item[$field]);
+        return ArrayHelper::getValue($item, $field) === null;
     }
 }
