+++ src/Reader/Iterable/FilterHandler/GreaterThanOrEqual.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Data\Reader\IterableFilterHandler;
+namespace Yiisoft\Data\Reader\Iterable\FilterHandler;
 
 use DateTimeInterface;
