+++ src/Reader/Iterable/FilterHandler/Group.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Data\Reader\IterableFilterHandler;
+namespace Yiisoft\Data\Reader\Iterable\FilterHandler;
 
 use InvalidArgumentException;
 use Yiisoft\Data\Reader\FilterAssertHelper;
-use Yiisoft\Data\Reader\IterableFilterHandlerInterface;
+use Yiisoft\Data\Reader\Iterable\IterableFilterHandlerInterface;
 
 use function array_shift;
 use function count;
