+++ tests/Paginator/OffsetPaginatorTest.php
@@ -8,7 +8,7 @@ use InvalidArgumentException;
 use Yiisoft\Data\Paginator\OffsetPaginator;
 use Yiisoft\Data\Paginator\PaginatorException;
 use Yiisoft\Data\Reader\CountableDataInterface;
-use Yiisoft\Data\Reader\IterableDataReader;
+use Yiisoft\Data\Reader\Iterable\IterableDataReader;
 use Yiisoft\Data\Reader\OffsetableDataInterface;
 use Yiisoft\Data\Reader\ReadableDataInterface;
 use Yiisoft\Data\Reader\Sort;
